oauth2_app/requirements.txt
fastapi>=0.111.0
bcrypt==4.0.1
uvicorn[standard]>=0.29.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
jinja2>=3.1.4
